# ENH/PERF: pyarrow timestamp & duration conversion consistency/performance